refactor: split script into multiple smaller ones
